+++ R/panel-dimplot.R
@@ -38,7 +38,7 @@ dimplot_ui <- function(id, data) {
                     c("None" = "", stats::setNames(cats, cats))),
         .scroll_aspect_input(ns))
     ),
-    .scroll_plot_area(ns, "460px", csv = TRUE)
+    .scroll_plot_area(ns, csv = TRUE)
   )
 }