+++ R/panel-pseudobulk.R
@@ -134,7 +134,7 @@ pseudobulk_de_ui <- function(id, data) {
             div(class = "scroll-plot-bar",
                 .scroll_size_slider(ns),
                 .scroll_dl_button(ns("png"), "PNG"), .scroll_dl_button(ns("pdf"), "PDF")),
-            .scroll_spin(plotOutput(ns("plot"), height = "520px")))))
+            .scroll_spin(plotOutput(ns("plot"), height = .SCROLL_PLOT_H)))))
   )
 }