+++ R/panel-spatial.R
@@ -97,7 +97,7 @@ spatial_ui <- function(id, data) {
             .scroll_size_slider(ns),
             .scroll_dl_button(ns("png"), "PNG"), .scroll_dl_button(ns("pdf"), "PDF"),
             .scroll_dl_button(ns("csv"), "CSV")),
-        .scroll_spin(plotOutput(ns("plot"), height = "560px",
+        .scroll_spin(plotOutput(ns("plot"), height = .SCROLL_PLOT_H,
             brush = shiny::brushOpts(ns("brush"), resetOnNew = TRUE),
             dblclick = ns("dblclick")))))
 }